src/backend/requirements.txt
scrapy 
beautifulsoup4 
pillow 
pypdf 
pdfplumber 
fastapi 
uvicorn 
pydantic 
python-dotenv 
pytesseract 
python-multipart 
starlette 
fastapi[standard]
axios 
numpy 
faiss-cpu 
langchain 
langchain_community 
langchain_text_splitters 
langchain_openai 
langchain_chroma 
gradio 
PyPDF2
python-docx